@@ -12,19 +12,19 @@
         "email": "dmitry.g.ivanov@gmail.com"
     }],
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "arvenil/ninja-mutex": "dev-master#82cbb2c",
-        "illuminate/console": "^10.0",
-        "illuminate/support": "^10.0"
+        "illuminate/console": "^11.0",
+        "illuminate/support": "^11.0"
     },
     "require-dev": {
         "ext-redis": "*",
         "ext-pdo_mysql": "*",
-        "phpunit/phpunit": "^10.5",
-        "mockery/mockery": "^1.5.1",
-        "orchestra/testbench": "^8.0",
-        "illuminated/testing-tools": "^10.0",
-        "predis/predis": "^2.0.2"
+        "phpunit/phpunit": "^11.3.6",
+        "mockery/mockery": "^1.6.10",
+        "orchestra/testbench": "^9.11.2",
+        "illuminated/testing-tools": "^11.0",
+        "predis/predis": "^2.3"
     },
     "autoload": {
         "psr-4": {
