@@ -15,11 +15,8 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
 
     /**
      * Resolve application Console Kernel implementation.
-     *
-     * @param \Illuminate\Foundation\Application $app
-     * @return void
      */
-    protected function resolveApplicationConsoleKernel($app)
+    protected function resolveApplicationConsoleKernel($app): void
     {
         $app->singleton(KernelContract::class, Kernel::class);
 
